iHome-python/requirements.txt
alembic==1.4.3
aliyun-python-sdk-core==2.13.28
aliyun-python-sdk-core-v3==2.13.11
altgraph==0.17
amqp==5.0.2
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
baidu-aip==2.2.18.0
billiard==3.6.3.0
blinker==1.4
cachelib==0.1.1
captcha==0.3
celery==5.0.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-formtools==2.2
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-redis-sessions==0.6.1
django-reversion==3.0.8
django-simple-captcha==0.5.12
django-tinymce==3.0.2
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-extensions==0.6.0
eventlet==0.29.1
Faker==4.1.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
html2text==2020.1.16
httplib2==0.18.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.6.1
incremental==17.5.0
itchat==1.3.10
itemadapter==0.1.0
itsdangerous==1.1.0
itypes==1.2.0
jieba==0.39
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lxml==4.5.1
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
mime==0.1.0
msgpack-python==0.5.6
mysql-connector-python==8.0.18
mysqlclient==1.4.6
numpy==1.19.4
parsel==1.6.0
pefile==2019.4.18
Pillow==7.1.2
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.4
PyDispatcher==2.0.5
pyflakes==2.2.0
pygame==1.9.6
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pypng==0.0.20
PyQRCode==1.2.1
pyStrich==0.8
python-alipay-sdk==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2017.3
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qiniu==7.3.0
qrcode==6.1
queuelib==1.5.0
raven==6.10.0
redis==3.5.3
requests==2.24.0
Scrapy==2.2.0
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
utils==1.0.1
vine==5.0.0
virtualenv==20.0.25
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.1
XlsxWriter==1.3.7
xlwt==1.3.0
yapf==0.30.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.0
